requirements.txt
python-dotenv
requests
gspread
google-auth
schedule